### 2024-09-25 #7279 @editorialbot
[PRE REVIEW]: PySHbundle: A Python tool for processing GRACE Gravimetry data into Global Surface Mass Change Datasets
**Submitting author:** @lsmvivek (<a href="http://orcid.org/0009-0000-7156-4450">Vivek Kumar Yadav</a>)
**Repository:** https://github.com/GESS-research-group/pyshbundle
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.0
