### 2024-09-25 #7279 @kthyng
I know this is a resubmission — please note here any important items from previous submission we should know (here is a link to the last one https://github.com/openjournals/joss-reviews/issues/7183).
