### 2024-09-26 293aac3
Delete LICENSE.md
