### 2024-09-26 bbad419
updated license and paper ackwgld
