### 2024-09-27 #7279 @lsmvivek
> Hi @lsmvivek and thanks for your submission! I am looking for some specific items to make sure your submission fits our requirements at a high level (not at the more detailed review level) before moving on to finding an editor or putting this on our waitlist if no relevant editors are available. I'll comment over time as I have a chance to go through them:
>
> * [ ]  [OSI-approved](https://opensource.org/licenses) license
> * [ ]   Installation instructions
> * [ ]   Documentation exists at all
> * [ ]  Tests of some sort
> * [ ]  Check paper (make sure compiles, read through, check for length [250-1000 words](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain))
> * [ ]   [Substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
> * [ ]  Clear research application
>
> In the meantime, please take a look at the comments above ⬆️ from the editorialbot to address any DOI, license, or paper issues if you're able (there may not be any), or suggest reviewers. For reviewers, please suggest 5 reviewers from the [database listed above](https://reviewers.joss.theoj.org/lookup) or your own ([non-conflicted](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)) extended network. Their github handles are most useful to receive but please don't use "@" to reference them since it will prematurely ping them.
>
> I know this is a resubmission — please note here any important items from previous submission we should know (here is a link to the last one #7183).

Hello @kthyng,
Indeed this is a resubmission. I went through the reviewer database and I suggest the following 5 people based on their work profile. Please let me know if anything else is needed.
Geodesy:
    busstoptaktik
    leouieda
    jiangyuinsar
Glaciers:
    JordiBolibar
    tgoelles
