### 2024-10-02 #7279 @kthyng
@lsmvivek
* docs don't seem to be working at https://mn5hk.github.io/pyshbundle/
