### 2024-10-02 ce2df0d
Update mkdocs.yml
