+++ docs/tutorials.md
@@ -2,9 +2,9 @@
 
 In order to make getting started with PySHBundle simple, we have curated a set of introductory tutorials in form of jupyter notebooks. Data for trying out this new tool is included in the repo. After installing and cloning the repo, go to the notebooks directory in order to find explainatory ipython jupyter notebooks. Simply activate the virtual environment and fire up these jupyter notebooks. The following notebooks explain the usage of the tool aswell as some crucial basics of Spherical Harmonics processing for Grace.
 
-1. [Introduction to Spherical Harmonics](https://github.com/mn5hk/pyshbundle/blob/main/notebooks/01_intro_spherical_harmonics_with_pyshbundle.ipynb)
-2. [Loading the data](https://github.com/mn5hk/pyshbundle/blob/main/notebooks/02_load_data.ipynb)
-3. [Visualizations](https://github.com/mn5hk/pyshbundle/blob/main/notebooks/03_visualizations.ipynb)
-4. [Terrestrial Water Storage (TWS) Time Series](https://github.com/mn5hk/pyshbundle/blob/main/notebooks/04_TWS_time_series.ipynb)
-5. [Tests and Validation notebook](https://github.com/mn5hk/pyshbundle/blob/main/notebooks/tests_validation_pyshbundle.ipynb)
+1. [Introduction to Spherical Harmonics](https://github.com/GESS-research-group/pyshbundle/blob/main/examples/01_intro_spherical_harmonics_with_pyshbundle.ipynb)
+2. [Loading the data](https://github.com/GESS-research-group/pyshbundle/blob/main/examples/02_load_data.ipynb)
+3. [Visualizations](https://github.com/GESS-research-group/pyshbundle/blob/main/examples/03_visualizations.ipynb)
+4. [Terrestrial Water Storage (TWS) Time Series](https://github.com/GESS-research-group/pyshbundle/blob/main/examples/04_TWSA_time_series.ipynb)
+5. [Tests and Validation notebook](https://github.com/GESS-research-group/pyshbundle/blob/main/examples/validation_pyshbundle.ipynb)
 
