### 2024-10-02 8de24c6
Update mkdocs.yml
