+++ mkdocs.yml
@@ -1,6 +1,6 @@
 site_name: pyshbundle
 
-site_url: https://abhimhamane.github.io/pyshbundle
+site_url: https://gess-research-group.github.io/pyshbundle/
 
 repo_url: https://github.com/GESS-research-group/pyshbundle
 
