+++ docs/installation.md
@@ -9,6 +9,14 @@ This is the preferred method to install pyshbundle, as it will always install th
 If you don't have [pip](https://pip.pypa.io) installed, this [Python installation guide](http://docs.python-guide.org/en/latest/starting/installation/) can guide you through the process.
 
 ```shell
+# clone the repository in order to access the notebooks and data
+$ git clone https://github.com/lsmvivek/pyshbundle.git
+$ pip install .
+
+
+# The package is available on pip but is BROKEN
+# Please avoid installing via pip till we fix that
+
 # creating a new virtual environment
 $ python3 -m venv <name-env>
 # activate the virtual environment environment
@@ -17,7 +25,7 @@ $ source </location-of-virt-env/name-env/bin/activate>
 $ pip install pyshbundle
 
 # clone the repository in order to access the notebooks and data
-$ git clone git@github.com:mn5hk/pyshbundle.git
+$ git clone https://github.com/lsmvivek/pyshbundle.git
 ```
 
 ## From the source - for Devs/Contributors
@@ -26,7 +34,7 @@ Developers can access the latest development branch and
 
 ```shell
 # clone the repo and fetch the dev branch
-$ git clone git@github.com:mn5hk/pyshbundle.git
+$ git clone https://github.com/lsmvivek/pyshbundle.git
 
 # creating a new virtual environment
 $ python3 -m venv <name-env>
