+++ docs/pyshbundle.md
@@ -4,9 +4,9 @@
   The module codes can be categorized into four categories:
 
   + [Load Data Modules](load_data.md)
-  + [convert data formats](convert_data_formats.md)
-  + [core functionality](core_functionality.md)
-  + [auxillary codes](auxillary_codes.md)
+  + [Convert data formats](convert_data_formats.md)
+  + [Core functionality](core_functionality.md)
+  + [Auxillary codes](auxillary_codes.md)
 
   ![Schematic diagram of code workflow](img/01_flowchart_20240327.jpg)
 
