### 2024-10-03 1fa53f6
Update shutils.py
