+++ pyshbundle/shutils.py
@@ -66,8 +66,9 @@ def plm(l: np.array, m:int, thetaRAD, nargin, nargout):
         number of input argument
     nargout : int
         number of output argument
+
     Returns
-    ----------
+    -------
     p : array
         plm fully normalized Legendre functions
     dp : array
@@ -197,8 +198,8 @@ def plm(l: np.array, m:int, thetaRAD, nargin, nargout):
 
   
 def secrecur(m, y):
-    """Helper Function for sectorial recursion.
-
+    """
+    Helper Function for sectorial recursion.
     This function computes the sectorial recursion for given parameters.
 
     Parameters
@@ -209,7 +210,7 @@ def secrecur(m, y):
         The input array for which the recursion is computed.
 
     Returns
-    ----------
+    -------
         numpy.ndarray: The result of the sectorial recursion.
     """
     if m == 0:
@@ -224,23 +225,23 @@ def secrecur(m, y):
 # % function for the l-recursion
 
 def lrecur(inn, x, m, lmax):
-    """Helper Function
+    """Helper function for recursion.
 
     Parameters
     ----------
     inn : int
-        _description_
+        Input value representing the initial condition.
     x : int
-        _description_
+        The current value for the recursion.
     m : int
-        _description_
+        Order of the recursion.
     lmax : int
-        _description_
+        Maximum value for recursion.
 
     Returns
-    ----------
-    inn : 
-        _description_
+    -------
+    inn : int
+        Updated value after performing the recursion based on parameters.
     """
     for ll in np.arange(int(m)+1,lmax+1,1):
        col   = ll - m+1			                                                # points to the next collumn of ptmp
@@ -263,15 +264,21 @@ def derivALF(inn, miin, plin, m, lmax):
 
     Parameters
     ----------
-    inn (numpy.ndarray) : _description_
-    miin (numpy.ndarray) : _description_
-    plin (numpy.ndarray) : _description_
-    m (int) : order of associated legendre functions
-    lmax (int) : maximum degree
+    inn : np.ndarray 
+        _description_
+    miin : np.ndarray
+        _description_
+    plin : np.ndarray
+        _description_
+    m : int 
+        order of associated legendre functions
+    lmax : int 
+        maximum degree
 
     Returns
     -------
-        numpy.ndarray : derivatives of the associated Legendre functions
+    inn : np.ndarray
+        derivatives of the associated Legendre functions
     """
     l = np.arange(m,lmax+2,1)
     if m == 0:
@@ -292,8 +299,8 @@ def derivALF(inn, miin, plin, m, lmax):
 
 
 def iplm(l, m:int, theRAD, dt=-9999):
-    """iplm Integrals of the fully normalized associated Legendre functions
-        over blocks for a selected order M. 
+    """
+    iplm Integrals of the fully normalized associated Legendre functions over blocks for a selected order M. 
 
     Parameters
     ----------
@@ -308,8 +315,8 @@ def iplm(l, m:int, theRAD, dt=-9999):
         integration block-size [rad] (scalar). Defaults to -9999.
     
     Returns
-    ----------
-    numpy.ndarray
+    -------
+    p : np.ndarray
         Matrix with integrated Legendre functions.
         Functions are integrated from theRAD(i)-dt/2 till theRAD(i)+dt/2.
         The matrix has length(TH) rows and length(L) columns, unless L 
@@ -443,7 +450,7 @@ def ispec(a,b = -9999):
         sine coefficients. Defaults to -9999.
 
     Returns
-    ----------
+    -------
         f (numpy.ndarray: **fill**
 
     See Also:
@@ -474,7 +481,6 @@ def ispec(a,b = -9999):
 
[…]
