### 2024-10-04 593b26d
updates to links of docs (mkdocs)
