+++ pyshbundle/shutils.py
@@ -1,8 +1,6 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 # Spherical Harmonics Utilities
 # Curator: Abhishek Mhamane
-
+# Updated: Vivek, 2024-10-04
 # - - - - - - - - - - - - - - 
 # License:
 #    This file is part of PySHbundle.
@@ -51,33 +49,23 @@ from pyshbundle import GRACEconstants as GC
 
 
 def plm(l: np.array, m:int, thetaRAD, nargin, nargout): 
-    """plm Fully normalized associated Legendre functions for a selected order M
+    """
+    Fully normalized associated Legendre functions for a selected order M.
 
-    Parameters
-    ----------
-    l : numpy.array
-        Degree, but not necessarily monotonic.
-            For l < m a vector of zeros will be returned.
-    m : int
-        order. If absent, m = 0 is assumed.
-    thetaRAD : numpy.array
-        Co-latitude in radians
-    nargin : int
-        number of input argument
-    nargout : int
-        number of output argument
+    Args:
+        l (numpy.array): Degree, but not necessarily monotonic. For l < m a vector of zeros will be returned.
+        m (int): Order. If absent, m = 0 is assumed.
+        thetaRAD (numpy.array): Co-latitude in radians.
+        nargin (int): Number of input arguments.
+        nargout (int): Number of output arguments.
 
-    Returns
-    -------
-    p : array
-        plm fully normalized Legendre functions
-    dp : array
-        first derivative of plm
-    ddp : array
-        second derivative of plm
-    
+    Returns:
+        (numpy.array): Fully normalized Legendre functions.
+        (numpy.array): First derivative of the Legendre functions.
+        (numpy.array): Second derivative of the Legendre functions.
+
     Author:
-        Vivek Kumar Yadav, Interdisciplinary Center for Water Research (ICWaR), Indian Institute of Science (IISc)
+        Vivek Kumar Yadav, Interdisciplinary Center for Water Research (ICWaR), Indian Institute of Science (IISc).
     """
 
     if  min(l.shape) != 1:
@@ -202,15 +190,11 @@ def secrecur(m, y):
     Helper Function for sectorial recursion.
     This function computes the sectorial recursion for given parameters.
 
-    Parameters
-    ----------
-    m : int
-        The order of the recursion.
-    y : numpy.ndarray
-        The input array for which the recursion is computed.
+    Args:
+        m (int): The order of the recursion.
+        y (numpy.ndarray): The input array for which the recursion is computed.
 
-    Returns
-    -------
+    Returns:
         numpy.ndarray: The result of the sectorial recursion.
     """
     if m == 0:
@@ -225,23 +209,17 @@ def secrecur(m, y):
 # % function for the l-recursion
 
 def lrecur(inn, x, m, lmax):
-    """Helper function for recursion.
+    """
+    Helper function for recursion.
 
-    Parameters
-    ----------
-    inn : int
-        Input value representing the initial condition.
-    x : int
-        The current value for the recursion.
-    m : int
-        Order of the recursion.
-    lmax : int
-        Maximum value for recursion.
+    Args:
+        inn (int): Input value representing the initial condition.
+        x (int): The current value for the recursion.
+        m (int): Order of the recursion.
+        lmax (int): Maximum value for recursion.
 
-    Returns
-    -------
-    inn : int
-        Updated value after performing the recursion based on parameters.
+    Returns:
+        (int): Updated value after performing the recursion based on parameters.
     """
     for ll in np.arange(int(m)+1,lmax+1,1):
[…]
