### 2024-10-04 #7279 @lsmvivek
> @lsmvivek
>
> * docs don't seem to be working at https://mn5hk.github.io/pyshbundle/

Please follow the updated link: https://gess-research-group.github.io/pyshbundle/
