### 2024-10-07 5b27f4c
Update pysh_core.py
