### 2024-10-07 f3f2fd9
Update pysh_core.py
