+++ faq/faq.md
@@ -1,0 +1,2 @@
+# FAQ
+Work in Progress...
