+++ pyshbundle/pyshbundle.md
@@ -1,0 +1,13 @@
+ 
+# Reference Mannual - PySHBundle
+
+  The module codes can be categorized into four categories:
+
+  + [Load Data Modules](load_data.md)
+  + [Convert data formats](convert_data_formats.md)
+  + [Core functionality](core_functionality.md)
+  + [Auxillary codes](auxillary_codes.md)
+
+  ![Schematic diagram of code workflow](img/pyshbundle_structure.png)
+
+  Navigate the Reference Manual based on the following schematic
