+++ acknowledgement/acknowledgement.md
@@ -1,0 +1,21 @@
+# Acknowledgement:
+
+The authors would like to thank Dr.-Ing. Markus Antoni and Clara Buetzler, Institute of Geodesy, University of Stuttgart, Germany, for early feedback. We are grateful for the financial support from IISc-ISRO Space Technology Cell for funding the project titled "Improving the spatial resolution of GRACE TWS for India using remote sensing datasets and modeling approach" under grant number STC0437. BDV would like to acknowledge the financial support from Science and Engineering Research Board, Government of India, under the grand agreement number SRG/2022/000625 for the MATRA project.
+
+Please note that PySHbundle has adapted the following code packages, both licensed under GNU General Public License
+
+   1. SHbundle: https://www.gis.uni-stuttgart.de/en/research/downloads/shbundle/
+
+   2. Downscaling GRACE Total Water Storage Change using Partial Least Squares Regression - https://springernature.figshare.com/collections/Downscaling_GRACE_Total_Water_Storage_Change_using_Partial_Least_Squares_Regression/5054564 
+
+
+# Key Papers Referred:
+   1. Vishwakarma, B. D., Horwath, M., Devaraju, B., Groh, A., & Sneeuw, N. (2017). 
+      A data‐driven approach for repairing the hydrological catchment signal damage 
+      due to filtering of GRACE products. Water Resources Research, 
+      53(11), 9824-9844. https://doi.org/10.1002/2017WR021150
+
+   2. Vishwakarma, B. D., Zhang, J., & Sneeuw, N. (2021). 
+      Downscaling GRACE total water storage change using 
+      partial least squares regression. Scientific data, 8(1), 95.
+      https://doi.org/10.1038/s41597-021-00862-6
