+++ auxillary_codes/auxillary_codes.md
@@ -1,0 +1,7 @@
+# Auxillary Codes
+
+The helper functions of the core functions are under the `shutils` script. 
+See [Core Functionality](core_functionality.md).
+
+
+::: pyshbundle.shutils

[…]
