### 2024-10-28 #7279 @lsmvivek
> @lsmvivek Did I already suggest giving a more descriptive title to this paper? It seems one already needs to know what this package does in order to know what the title means. I struggle with this whenever I try to search for an editor for the submission.

Dear @kthyng
I have updated the paper title.
It clearly states the objective of the software, data it works with and the expected output.
I have also added more keywords (Geodesy, Gravimetry, Terrestrial Water Storage).

I thank you for suggestion, please let me know if any other changes are needed.
