### 2024-11-27 #7279 @MMesch
@hugoledoux Yes I would be willing to do this depending on exact criteria and timeline. I wrote you a DM on Mastodon for further contact details.
