### 2024-12-02 #7279 @hugoledoux
> @hugoledoux Yes I would be willing to do this depending on exact criteria and timeline. I wrote you a DM on Mastodon for further contact details.

@MMesch I sent you an email last week, have you received it?
