### 2024-12-03 #7550 @editorialbot
[REVIEW]: PySHbundle: A Python software to convert GRACE Spherical Harmonic Coefficients to gridded mass change fields
**Submitting author:** @lsmvivek (<a href="http://orcid.org/0009-0000-7156-4450">Vivek Kumar Yadav</a>)
**Repository:** https://github.com/GESS-research-group/pyshbundle
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.0
