### 2024-12-03 #7279 @hugoledoux
sorry for the delay @lsmvivek, we now have 2 very suitable reviewers and it's started at https://github.com/openjournals/joss-reviews/issues/7550
