### 2025-01-09 #7550 @hugoledoux
@MMesch @tsutterley just a reminder so that you don't forget this review
