### 2025-01-09 GESS-research-group/pyshbundle#132 @tsutterley
`theory.md` fixes
(this and following issues contributed as part of JOSS review)

Overall looks pretty great. Just a couple of little fixes.

- Level-0: ["corresponding" and "solutions, space"](
https://github.com/GESS-research-group/pyshbundle/blob/6cc3f6f1bc2cb35e3271ce18c21d27d92364b73a/docs/theory.md?plain=1#L160C502-L160C514)
- Level-1: want to mention the laser ranging data for GRACE-FO? Might also want to mention the auxiliary Level-1b data produced by GFZ for removing non-tidal atmospheric and oceanic variation.
- Level-2: JPL produces Level-2 data now at all times and not just for validation.
- Level-3: also includes the scaled gridded products produced by JPL. might want to just rename this to "spatial fields"
- ["gridded" and " Anomaly (TWSA)"](https://github.com/GESS-research-group/pyshbundle/blob/6cc3f6f1bc2cb35e3271ce18c21d27d92364b73a/docs/theory.md?plain=1#L175C1-L176C1)
- I think [this line](https://github.com/GESS-research-group/pyshbundle/blob/6cc3f6f1bc2cb35e3271ce18c21d27d92364b73a/docs/theory.md?plain=1#L178C11-L178C19) should be in the next section about formats
- should be ["triangular"](https://github.com/GESS-research-group/pyshbundle/blob/6cc3f6f1bc2cb35e3271ce18c21d27d92364b73a/docs/theory.md?plain=1#L215)
