### 2025-01-09 GESS-research-group/pyshbundle#133 @tsutterley
`installation.md` 
should be able to use `python3 -m pip git+https://github.com/GESS-research-group/pyshbundle.git` to install directly from the git repo right?

also I think using `python setup.py sdist` is [currently frowned upon](https://packaging.python.org/en/latest/guides/distributing-packages-using-setuptools/)
