### 2025-01-09 GESS-research-group/pyshbundle#134 @tsutterley
`01_intro_spherical_harmonics_with_pyshbundle.ipynb`
- for the $P_{3,0}$ plot, $P_{2,0}$ is accidentally plotted alongside the plot of $P_{3,0}*P_{3,0}$
- minor thing but the variable names and plot labels in the $P_{3,1}*P_{3,2}$ plot and the $P_{3,1}*P_{4,2}$ plot should be fixed
