### 2025-01-18 #7550 @tsutterley
Some overall comments for `PySHbundle` for openjournals/joss-reviews#7550:

I put together some issues (GESS-research-group/pyshbundle#132 GESS-research-group/pyshbundle#133 GESS-research-group/pyshbundle#134) that I found looking over the documentation. I still need to build an environment to run the programs themselves. I thought that the [Theoretical Background](https://gess-research-group.github.io/pyshbundle/theory/) page was really good.

Line-by-Line
------------

- [**Line 18**](https://github.com/GESS-research-group/pyshbundle/blob/6cc3f6f1bc2cb35e3271ce18c21d27d92364b73a/paper.md?plain=1#L69) All NASA satellite products have processing levels (see [NASA Earthdata learn site](https://www.earthdata.nasa.gov/learn/earth-observation-data-basics/data-processing-levels))
- [**Line 18**](https://github.com/GESS-research-group/pyshbundle/blob/6cc3f6f1bc2cb35e3271ce18c21d27d92364b73a/paper.md?plain=1#L69) I think "GRACE has transformed Geosciences" is a little strong. I might also rephrase the earlier portion of the sentence as "One of the major mechanisms of short term mass transport is the redistribution of water".
- [**Line 23**](https://github.com/GESS-research-group/pyshbundle/blob/6cc3f6f1bc2cb35e3271ce18c21d27d92364b73a/paper.md?plain=1#L69) "The L2 spherical harmonic data are typically noisy, which necessitates the use of spectral filtering. The data also have to be corrected for known artifacts and contaminating geophysical signals, such as solid Earth processes in the case of isolating TWSA."
- [**Line 26**](https://github.com/GESS-research-group/pyshbundle/blob/6cc3f6f1bc2cb35e3271ce18c21d27d92364b73a/paper.md?plain=1#L69) probably could be "*most* L3 users"
- [**Line 27**](https://github.com/GESS-research-group/pyshbundle/blob/6cc3f6f1bc2cb35e3271ce18c21d27d92364b73a/paper.md?plain=1#L69) "The majority" and "most often use the off-the-shelf L3 products"
- [**Line 30**](https://github.com/GESS-research-group/pyshbundle/blob/6cc3f6f1bc2cb35e3271ce18c21d27d92364b73a/paper.md?plain=1#L69) "called PySHbundle, was developed to ease the conversion of GRACE L2"
- [**Line 34**](https://github.com/GESS-research-group/pyshbundle/blob/6cc3f6f1bc2cb35e3271ce18c21d27d92364b73a/paper.md?plain=1#L73) "The NASA/DLR GRACE and NASA/GFZ GRACE-FO twin satellite missions measure changes in the Earth's gravitational field by measuring their inter-satellite distance. Changes in the local gravity field affect the orbit of each satellite, which is recorded with the onboard ranging system."
- [**Line 42**](https://github.com/GESS-research-group/pyshbundle/blob/6cc3f6f1bc2cb35e3271ce18c21d27d92364b73a/paper.md?plain=1#L75) this sentence starts in the middle??
- [**Line 42**](https://github.com/GESS-research-group/pyshbundle/blob/6cc3f6f1bc2cb35e3271ce18c21d27d92364b73a/paper.md?plain=1#L75) Should note that the JPL mascons are created directly from the Level-1B satellite ranges
- [**Line 47**](https://github.com/GESS-research-group/pyshbundle/blob/6cc3f6f1bc2cb35e3271ce18c21d27d92364b73a/paper.md?plain=1#L77) Might want to mention [`frommle2`](https://github.com/strawpants/frommle2) from @strawpants instead of his (`GRACE-filter`)[https://github.com/strawpants/GRACE-filter] set of programs
- [**Line 51**](https://github.com/GESS-research-group/pyshbundle/blob/6cc3f6f1bc2cb35e3271ce18c21d27d92364b73a/paper.md?plain=1#L77) "Spherical"
- [**Line 52**](https://github.com/GESS-research-group/pyshbundle/blob/6cc3f6f1bc2cb35e3271ce18c21d27d92364b73a/paper.md?plain=1#L77) "while the latest release was from 2021"
- [**Line 71**](https://github.com/GESS-research-group/pyshbundle/blob/6cc3f6f1bc2cb35e3271ce18c21d27d92364b73a/paper.md?plain=1#L94) Degree 1 isn't just poorly measured, it is inherently zero in a center-of-mass (CM) reference frame. Also "JPl"&rarr;"JPL" and "ITGZ"&rarr;"GFZ"
