### 2025-01-22 #7550 @lsmvivek
@hugoledoux @tsutterley
I have addressed:
- The above issues
- The issues raised at: [GESS-research-group#132](https://github.com/GESS-research-group/pyshbundle/issues/132), [GESS-research-group#134](https://github.com/GESS-research-group/pyshbundle/issues/134)

All changes are in [joss-review-7550](https://github.com/GESS-research-group/pyshbundle/tree/joss-review-7550) branch
