### 2025-01-22 #7550 @strawpants
Just noticed that [frommle2](https://github.com/strawpants/frommle2) was mentioned in this thread. Most of the work has now moved to [shxarray](https://github.com/ITC-Water-Resources/shxarray). So if you decide to mention it it's better to mention shxarray, because it's actively maintained.

thanks,
Roelof
