### 2025-01-23 #7550 @hugoledoux
> All changes are in [joss-review-7550](https://github.com/GESS-research-group/pyshbundle/tree/joss-review-7550) branch

if that branch is only temporary and you'll merge soon it's fine. But if you want our bot to build the paper from that specific branch then I can change that in the details above this page. Let me know.
