### 2025-01-24 #7550 @hugoledoux
okay all is fine, make some changes and then merge them.

I was just warning you that if you modify the paper and ask our bot to compile the PDF it will use the main branch.
