### 2025-01-27 #7550 @lsmvivek
> I was just warning you that if you modify the paper and ask our bot to compile the PDF it will use the main branch.

I have merged the branch [joss-review-7550](https://github.com/GESS-research-group/pyshbundle/tree/joss-review-7550) branch, where I implemented the [changes](https://github.com/openjournals/joss-reviews/issues/7550#issuecomment-2606907191), into the main branch.
