### 2025-02-11 #7550 @hugoledoux
@MMesch just a small reminder not to forget this review
