### 2025-02-20 GESS-research-group/pyshbundle#138 @MMesch
joss review 7550
Hi all,

Thanks for making this code available! I have started going through the article and the code and will open issues as I'm working through it. I already opened a PR #137 with some small suggestions for paper.md. A few more that I couldn't address immediately are below:

More superficial remarks:

- a short explanation or reference for JPL MASCONS would be nice;
- "to obtain Level 2 (`L2`) GRACE Spherical Harmonics data" maybe add whether these are the actual coefficients;

Deeper content remarks:

- as it stands, the paper is very terse and seems to be mostly meant as a reference that can be cited. That's fine if it's OK for JOSS, but my feeling is that for many readers a little more context would be useful. I also think it can be accomplished without making the article too long for the initiated. Personally, I would suggest using modern writing tools such as LLMs, of course under guidance, validation and injection of knowledge, to achieve this.
- add a paragraph about what differentiates pyshbundle from other options for spherical harmonics analysis that were cited.
- incorporating a short code example would be nice as well.
- adding a paragraph about how the software was validated would be useful. For example, with benchmarks against other results or other types of tests.
