### 2025-02-28 GESS-research-group/pyshbundle#138 @lsmvivek
> More superficial remarks:
>
> * a short explanation or reference for JPL MASCONS would be nice;
- Reference added
> * "to obtain Level 2 (`L2`) GRACE Spherical Harmonics data" maybe add whether these are the actual coefficients;
- Explicitly mentioned the solutions
> Deeper content remarks:
>
> * as it stands, the paper is very terse and seems to be mostly meant as a reference that can be cited. That's fine if it's OK for JOSS, but my feeling is that for many readers a little more context would be useful. I also think it can be accomplished without making the article too long for the initiated. Personally, I would suggest using modern writing tools such as LLMs, of course under guidance, validation and injection of knowledge, to achieve this.
- todo
> * add a paragraph about what differentiates pyshbundle from other options for spherical harmonics analysis that were cited.
- todo
> * incorporating a short code example would be nice as well.
- Examples already present in `examples/` and mentioned under _Implementation: 4. shutils_ in paper
> * adding a paragraph about how the software was validated would be useful. For example, with benchmarks against other results or other types of tests.
- todo
