### 2025-05-07 #7550 @lsmvivek
The suggested changes and corrections have been implemented in:

@tsutterley
- https://github.com/openjournals/joss-reviews/issues/7550#issuecomment-2606907191

@MMesch
- https://github.com/GESS-research-group/pyshbundle/issues/138#issuecomment-2689933939

I request the reviewers and editor to please provide further comments. @hugoledoux

Apologies for very long delay, I had to move continents for my work.
