### 2025-06-20 #7550 @hugoledoux
@MMesch @tsutterley Could you please pick this up again and continue the checklist based on the improvements made?
