### 2025-08-29 #7550 @hugoledoux
@MMesch @tsutterley the summer has ended ;) Another reminder that it would be nice to finish this review
