### 2026-02-27 #7550 @lsmvivek
Dear reviewer @MMesch and editor @hugoledoux,

This is a kind request to bring your attention towards the review of our paper.

PySHbundle, was written to ease the access of beginner and intermediate level users working on satellite geodesy data.

We translated the legacy SHbundle written in MATLAB into a widely used and open source programming language.

It has been a long time since we addressed the comments by reviewers.

Completed in the pre-Al generated code era, the code and the paper are result of many months spent by me and my co- authors @mn5hk, @WhythiskolaveriD, @abhimhamane and others. We took immense efforts in understanding the legacy code, carefully translating it into python given its different indexing styles, and rigorous validation at each level of mathematical operation applied to the data (apart from the validation script in the codebase).

Reviewer #1 Tyler Sutterley have completed their review process and accepted our changes.

We would benefit from your review comments and would be very grateful to complete the review process.

Further, we believe the larger Geodesy and Hydrology community would benefit from the review process. We plan to include PySHbundle in the coming editions of the annual GRACE HACKWEEK.

We sincerely hope you can help improve our code and the paper.

Yours truly,

Vivek
