### 2026-03-22 #7550 @MMesch
Sorry for the long wait and thanks for contributing this library to the open source world. I used a fresh commit for my review below. Some of the bullet points are comments and notes on what I did, others questions. Excusing in advance if I misunderstood anything or didn't find it in the code. Looking forward to getting this completely done.

---

_I'm inlining comments as I'm progressing through the review_

- [x] I confirm that I have read the JOSS conflict of interest (COI) policy
https://github.com/openjournals/joss/blob/master/COI.md and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.

- [x] I confirm that I read and will adhere to the JOSS code of conduct
https://joss.theoj.org/about#code_of_conduct.

- [x] Repository: Is the source code for this software available at the
https://github.com/GESS-research-group/pyshbundle?
  - a priori yes but completeness of the available source code will be checked later
  - this check could probably be automated by JOSS

- [x] License: Does the repository contain a plain-text LICENSE or COPYING file with the contents of an OSI approved https://opensource.org/licenses/alphabetical software license?
  - GPL
  - this check could probably be automated by JOSS

- [x] Contribution and authorship: Has the submitting author (@lsmvivek) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
  - a priori yes

- [x] Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines
https://joss.readthedocs.io/en/pre-2026-submission/submitting.html#substantial-scholarly-effort
  - [x] As a rule of thumb, JOSS’ minimum allowable contribution should represent not less than three months of work for an individual. Some factors that may be considered by editors and reviewers when judging effort include:
  - [x] Age of software (is this a well-established software project) / length of commit history.
      - first commit in Git Repo in 2022
      - regular contributions since 2022.
      - The original SHBundle code seems to be from 1994 as stated in the paper.
  - [x] Number of commits.
      - 763 commits total
  - [x] Number of authors.
      - three main contributors
  - [x] Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged those under 300 LOC will be desk rejected.
      - 4176
  - [x] Whether the software has already been cited in academic papers.
      - maybe not in the form of the python wrapper, at least I can't find any references but that's not unusual if this is the first article. The original software SHBundle has been cited. [Google Scholar shows 44 results](https://scholar.google.com/scholar?q=shbundle%20stuttgart&btnG=Search&as_sdt=800000000001&as_sdtp=on) for the search term "shbundle stuttgart" and some of them at least reference the "parent" of this software.
  - [x] Whether the software is sufficiently useful that it is likely to be cited by your peer group.
      - Not sure *my* peer group is appropriate here. But it does seem to be something useful and citable by the target users.

- [x] Data sharing: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.
  - no original data

- [x] Reproducibility: If the paper contains original results, results are entirely
reproducible by reviewers. If the paper contains no original results, please check this item.
  - I was able to reproduce test results, and the example notebooks (the latter high level since I didn't compare exactly the numbers on every plot).

- [x] Human and animal research: If the paper contains original data research on humans subjects or animals, does it comply with JOSS's human participants research policy and/or animal research policy
https://joss.readthedocs.io/en/pre-2026-submission/policies.html?highlight=animal#joss-policies? If the paper contains no such data, please check this item.
  - doesn't apply

  - Installation from PyPi [marked 2-years ago as BROKEN in the README](https://github.com/GESS-research-group/pyshbundle/blame/main/README.md#L26) and in [docs/installation.md](https://github.com/GESS-research-group/pyshbundle/blame/main/docs/installation.md#L17).
  - git clone install instructions are present in README.
  -  I am using an installation on NixOS with uv2nix: `nix shell nixpkgs#uv -c uv lock`, then copy flake.nix and flake.lock with uv2nix code [can provide in case interesting] in the folder, git add those flake files and the generated uv.lock, and run `nix develop`. This works.
- [x] Testing (minor, *I added this section, hope that's OK*)(*fixed, test run, see below*):
  - Unclear how to run tests since it seems not to be in the README. Using `python -m unittest tests.test_pyshbundle` as in CI workflow. Probably a good idea to migrate to something like pytest for easier examination of tests.
  - First test run failed because pkg_resources API was used throughout the code. From https://setuptools.pypa.io/en/latest/pkg_resources.html: "As of setuptools 81.0.0, the pkg_resources API is deprecated and scheduled for removal around 2025-11-30.". It seems that some non-uv, non-nix systems *do* provide setuptools even if not specified, and the removal was recent, so maybe this ran elsewhere. To overcome this, I decided to migrate the code from pkg_resources to importlib. An alternative could be to add downgraded setuptools (\<70) to dependencies but that would only be a temporary fix, the code already specifies Python >= 3.9 which should have importlib.
  - Afterwards tests were running albeit with
     - lots of warnings: `SyntaxWarning: "\S" is an invalid escape sequence. Such sequences will not work in the future. Did you mean "\\S"? A raw string is also an option.`, `RuntimeWarning: The global interpreter lock (GIL) has been enabled to load module 'cftime._cftime', which has not declared that it can run safely without the GIL. To override this behavior and keep the GIL disabled (at your own risk), run with PYTHON_GIL=0 or -Xgil=0.`.
     - one failure: `validation_pyshbundle.py", line 122, in validation_pyshbundle raise ValueError('Gridwise RMSE is greater than 1e-3')`. Added breakpoint to this test to find out more. Got very large values with major differences. Unsure what the issue is. The following is what I did for debugging:
       - Checking CI runs. Test seem to run on GitHub Actions CI but only for macos. Conda is used to setup the environment from environment.yml which means setuptools is included (explaining our prior exception). Then pip is used to install the package itself.
       - Moving to pixi in NixOS FHS user env to get closer to conda. Running `pixi init --import environment.yml`, then `pixi install`, then `pixi run python -m unittest tests.test_pyshbundle`. Same exception with exactly the same mismatching values. Unlikely to be a package manager issue.
       - But testing with conda (micromamba) also in NixOS FHS as third option to be sure: `micromamba create -f environment.yml -n pysh`, `micromamba activate pysh`, `pip install .`, `python -m unittest tests.test_pyshbundle`. Also trying to run
   with downgraded setuptools `pip install 'setuptools<70'` but exactly the same error.
       - conclusion for now: something is really wrong, maybe the matlab verification file is stale. Unclear though why GH CI seems to run successfully.
       - digging deeper it seems that this the failing test is comparing a computed matrix which has months as one dimension with a similar matrix computed in matlab. The computed matrix has *more* months than the matlab reference, no wonder there are differences. Speculating that maybe months were added since the test ran and truncating to matlab matrix shape. Still fails. Trying correlation and ... both matrices are highly correlated so this is a legit comparison but relative error is still 3% and absolute much higher than 1e-3. Xcorr shows that lag0 is best so that doesn't seem the issue either. Comparing month by month gets us further. Somehow 3 of the 196 months (105: 2011-10, 149: 2017-05, 155: 2019-01) are responsible for all the RMSE mismatch, producing values from 12 - 31. All other 193 have an RMSE of 2e-6 and passing the test. Conclusion: The code seems to work but there are some problematic months for whatever reason.
  - starting notebooks, `jupyter lab examples` starts flawlessly. All 5 provided notebooks run without any issue and seem to give high level meaningful results that resemble the included plots. I didn't verify every plot in detail though and had no time to go through all the physical statements made (yet). But from that high level perspective these notebooks seem thorough and really interesting.
  - inclined to say yes without having checked every use in detail. But the tests, including the notebook all run, meaningful spherical harmonics appear and can be used in various ways, notably to transform provided spherical harmonics coefficients to spatial domain with many ways to configure that transformation. Validation of the time series with respect to matlab works in 193 / 196 months and it is likely that the 3 failing months are a detail (to be confirmed).
- [x] Performance: If there are any performance claims of the software, have they
been confirmed? (If there are no claims, please check off this item.)
  - I am unsure how important performance optimizations are for the use cases that this library targets. It doesn't seem to be a major focus of the publication, and I didn't encounter any astonishing performance issues when running the example notebooks although that doesn't necessarily mean much. In case it is nevertheless a core addition it would be nice to add a benchmark but I'll check off this item.

  - This is clearly stated. A library to simplify the L2 → L3 transformation of spherical harmonics to spatial data, while exposing fine grained control how this transformation happens. Target researchers and students.
- [x] Installation instructions (*fixed with minor additional recommendations, see below*): Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
  - Pyproject.toml, requirements.txt and environment.yml are present with dependencies (covering pip, conda, uv and more ecosystems)
  - But this could be enhanced by updating pypi or other registry upload.
  - A lock file with pinned dependencies could be committed to enhance reproducibility.
  - It would also be nice to state how tests should be run to verify local installs.
  - Yes, 5 nice and detailed notebooks for high level end-to-end usage.
- [x] Functionality documentation: Is the core functionality of the software
documented to a satisfactory level (e.g., API method documentation)?
  - Checking off although this could be improved:
    - lots of unit tests that can be examined for detailed usage — great.
    - most functions have docstrings with Google style format. Most parameters
      are explained in these docstrings.
    - But I am not completely sure whether the docstring content and function signatures are actually validated. Are these all up to date? It seems these are *not* tested, no linting. In case the software continuous to evolve linter, formatter, type checker, docstring validator could be useful to verify this documentation.
    - documentation is built in a website hosted on GH pages. https://gess-research-group.github.io/pyshbundle/ .
    - some type hints are there although more could help to see what functions take.
    - variable names could be a bit more descriptive.
- [x] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
  - Yes, lots of unit tests.
  - As noted before, apparently no linting, type checking etc which could catch some extra issues.
- [x] Community guidelines: Are there clear guidelines for third parties wishing to
  1. Contribute to the software
  2. Report issues or problems with the software
  3. Seek support
  - covered in the README.

- [x] Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
  - Yes.
  - it seems a superfluous `=======` got into this? Maybe a git merge remnant?
- [x] A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?
  - Yes.
- [x] State of the field: Do the authors describe how this software compares to other commonly-used packages?
  - Extensive links to other packages are there
  - Differentiator not there (*fixed, see below*). Would be nice to add content about this. I am not sure how much JOSS requires and I am aware that this is real effort since you need to dig through other codes, and sometimes minor differences in integration are decisive in software engineering. The Statement of Need section may somewhat address this, stating compatibility with MATLAB codes, modular design and accessibility, but a more direct comparison to other tools would be nice because there is definitely overlap.
- [x] Quality of writing: Is the paper well written (i.e., it does not require
editing for structure, language, or writing quality)?
  - Yes, generally well written.
- [x] References: Is the list of references complete, and is everything cited
appropriately that should be cited (e.g., papers, datasets, software)? Do
references in the text use the proper citation syntax
https://pandoc.org/MANUAL.html#extension-citations?
  - minor issues (*fixed, see below*):
    - @watkins2015mascons may need [] around
    - some doi fields are empty

### Random

- Pure Python codebase. Matlab for validation included.
- would be nice to check whether this runs (at least partly) in emscripten-forge and jupyterlite
- Cybersecurity:
  - `bandit -r pyshtools` (checking for exec, eval, subprocess, pickle, passwords, SQL inj) returns no concerns (4 low severity findings).
  - `pip-audit` also found no vulnerabilities in dependencies
  - `trivy fs .` does find vulnerabilities in deps, some of them high severity and now fixed, but most seem not important when this library is used in the target research setting. Could be checked systematically, although I think GitHub also does basic checks for it.

#### Repository

- The GitHub About section "Convert SH bundle Matlab codes to python" is not very evocative and maybe even a bit misleading. I suggest updating to something closer to the README.md title. → Done.
