### 2026-04-04 #7550 @lsmvivek
Dear reviewer @MMesch,

Thank you for providing your comments. Based on your comments we have removed some deprecated dependencies, added type annotations and implemented tests with `pytest`. Please see the description below of the changes we made and our response. We hope to hear from you soon and complete the review process.

All changes are made in main branch: [main](https://github.com/GESS-research-group/pyshbundle)

Functionality

- Replaced python `unittests` with `pytest`.
- As implemented by you, we replaced the deprecated `pkg_resources` with `importlib`.
- Fixed dangling "\S" warnings.
- Confirm the failure of `validation_pyshbundle.py` was due to corrupted input data. The three months showed high error in our testing as well. We removed the corrupted files and reduced testing for full GRACE record to 60 months. Ignoring the GRACE data gaps, this allows testing it for 5 years of monthly data, making it a sufficient test of the accuracy of the software.
- Reduced size also allows meeting the pip limitation of 100MB.
- All tests passing on 2 Python versions × 3 operating systems = 6 combinations. Python versions 3.9 and 3.12. OS windows-latest, macOS-latest and Ubuntu-latest, see CI build-and-test workflow.
- Removed redundancies from pyproject.toml, environment.yml, requirements.txt and setup.py. Installation should work properly without any errors, across all platforms.

Documentation
- As described above, installation files updated.
- PyPi installation fixed. `pip install pyshbundle`.
- Lock file would be overly restrictive for a library. Besides `pyproject.toml` should be sufficient for reproducibility.
- Full description of testing implemented with `pytest` for users in README and in paper.
- Updated code with proper formatting, linting, added missing docstrings to some functions and added type annotations, throughout the codebase. Applied ruff format across all source files to enforce consistent code style. Further, configured `mypy` in `pyproject.toml` and ran it across the codebase. Together, these will ensure good continuous development standards.
- at this point we avoided changing variable names as most of the variables are using standard notations like P, l and m accordingly to their mathematical formulations.

Software paper
- Added testing description with `pytest`: "The accuracy of the core capabilities of `pyshbundle` was validated against the MATLAB `SHbundle` software using 60 months of JPL GRACE RL06 Level-2 data. The gridwise Root Mean Square Error (RMSE) between the computed and reference Terrestrial Water Storage fields was below $10^{-3}$ mm equivalent water height, confirming numerical equivalence between the two implementations".
- Removed superfluous `=======`.
- State of the field: Do the authors describe how this software compares to other commonly-used packages?
We describe its usage under the Statement of need. I forgot to merge the added text from the review branch into the main, hence it would not have appeared when you read it. Pardon, I have merged it now. The description is: "While excellent tools exist (above) for Spherical Harmonic operation, PySHbundle provides familiarity of existing and beginner level users by focusing on GRACE applications, and translating the legacy software, SHbundle." Further, a comprehensive feature-by-feature comparison with all related packages is beyond the scope of a software description paper.

- References
   - Corrected format for @watkins2015mascons. Thank you for pointing it out.
   - Removed empty DOI fields where no DOI was available.

Repository
- Updated the GitHub About section to "A Python implementation of GRACE Spherical Harmonics Synthesis"

- emscripten-forge and jupyterlite: Implementing this is beyond my technical capability and out of scope of our work.
