### 2026-04-04 #7550 @MMesch
Dear @lsmvivek,

Thank you for the comprehensive updates.

## I checked

- pytest migration with clear README documentation
- pkg_resources → importlib migration
- \S escape warnings fixed
- Test failures resolved (60 months, RMSE < 1e-3)
- Type annotations added to function signatures
- ruff formatting checks
- Paper artifacts (=======) removed
- Citation format and empty DOI fields fixed

## Minor Recommended Actions Left

*I won't insist if there are strong opinions but I think it would be useful to do the small following changes:*

- Add dev dependencies to pyproject.toml: `pytest`, `ruff`, and `mypy` are configured but not installed. E.g. add:
  ```toml
  [project.optional-dependencies]
  dev = ["pytest", "pytest-cov", "ruff", "mypy"]
  ```
  so that they will be picked up by package managers relying on pyproject.toml.
- The author is right that a library should *not* pin every dependency in pyproject.toml because that would restrict usage more than necessary. But a lock file should ideally *still* be in the repository to allow reproducing the provided tests and notebooks with the exact same versions. This lock file can be excluded from package distribution.
- Minor cleanup:
  - Remove orphan `</p>` tag from paper.md line 117
  - Add to .gitignore: `__pycache__/`, `*.pyc`, `.ipynb_checkpoints/` and maybe more. There are standard lists for .gitignore that can be used.

## Non-blocking

- Type checking: `mypy pyshbundle/` produces 49 errors (mostly numpy type unions which may be a bit more complex to solve). Acceptable but could improve over time to fully typecheck (even if it means selectively ignoring) and to include mypy in the CI.

## Noted

- Variable names using mathematical notation OK.
- State of field comparison (line 86) is minimal but OK. I can understand that it's out of scope.
- emscripten/jupyterlite: acknowledged beyond scope this was just a thought/idea.

I think the above recommendations can be addressed independently as the author sees fit. No further review is needed from my end but referring to @hugoledoux for this.
