### 2026-04-12 #7550 @lsmvivek
Dear @MMesch,

Thank you for your quick reply and the suggestions.
We have implemented your suggestions and merged all changes to [main](https://github.com/GESS-research-group/pyshbundle). As suggested:
- Added dev dependencies in pyproject.toml file.
- Added lock files with requirement-lock.txt and environment-lock.yml.
- Removed orphan `</p>` tag from paper.md file.
- Added `__pycache__/`, `*.py[co]`, `.ipynb_checkpoints/` to the .gitignore file.

Non-blocking
As you correctly pointed out the errors are due to many instances of numpy type unions and will need selective ignoring. It is much more complex than our understanding. Despite these, at the current point in time, the type annotations should be are comprehensive and should be sufficient for any user looking to understand the code.

We thank you again for your comments and suggestions. Hoping to complete the review soon process soon.
