### 2026-04-17 #7550 @hugoledoux
I am glad that this review has (almost) come to an end. It took some time but we got there and the software/docs are surely better now.

I have proof-read the paper and I thought that several sentences and words were not optimal and consistent, so I have modified where I saw fit. Please check the PR and accept unless you think I changed the meaning a sentence: https://github.com/GESS-research-group/pyshbundle/pull/147

The capitalisation of the title was a mix, so I put all in small letters except words that ought to be capitalised.

One thing that is still to fix: Nagesh Kumar Dasika has ORCID 0009-0006-9323-1191, which is identical to Tsungrojungla Walling's (lines 27/34).

And another: the DOI that is reported just above
