### 2026-04-20 #7550 @lsmvivek
Dear Editor @hugoledoux,

Thank for suggestions on the paper. I have checked and merged the changes without further edits.

I have corrected the ORCID of co-author Nagesh, thank you for pointing it out. And finally corrected the DOI of `watkins2015mascons`.

I hope the paper and software are now fully ready for final decision. Thank you for your suggestions, edits, efforts and paitence with our manuscript.

Checking the references below.
