### 2026-04-21 #7550 @lsmvivek
Dear Editor @hugoledoux,

We have implemented the tagged releasing and archiving. Please see below:

- Made a tagged release of `pyshbundle` with tag `v0.3.0`: [pyshbundle v0.3.0](https://github.com/GESS-research-group/pyshbundle/releases/tag/v0.3.0)
- Archived the reviewed software in Zenodo.
- Zenodo archive has the correct Metadata. Title same as paper, authors same as paper with ORCID.
- DOI of Zenodo archive: 10.5281/zenodo.19682215 ([10.5281/zenodo.19682215](https://zenodo.org/records/19682215))
