### 2026-04-22 #7550 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
